Ensure timestamp index can be monotonic
Pandas requires the index used when computing a rolling avg function to be monotonic.  Since the initial list of files in your populating of the panads dataframe makes no guarantee of order (especially on high throughput filesystems like XFS), the index needs to be sorted so when the group selected rolling avg is run it won't error for being non-monotonic (This happened to me, hence this fix).

Cheers!